rule_engine/requirements.txt
# requirements.txt
paho-mqtt
